+++ src/config/dictionary/generics/sources.mjs
@@ -7,3 +7,10 @@ export const SOURCE_CATEGORIES = {
     16, 9999999,
   ],
 };
+
+export const SOURCE_DATA = {
+  spellDescriptions2014: [150, 151, 152, 154, 155, 156, 158, 9999999],
+  spellDescriptions2024: [145, 146, 147, 148, 149, 153, 157],
+  is2014: [150, 151, 152, 154, 155, 156, 158, 9999999],
+  is2024: [145, 146, 147, 148, 149, 153, 157],
+};