+++ src/parser/race/DDBRace.js
@@ -84,9 +84,15 @@ export default class DDBRace {
   constructor(ddbData, race, compendiumRacialTraits, isGeneric = false) {
     this.ddbData = ddbData;
     this.race = race;
-    this.is2014 = this.race.isLegacy
-      && this.race.sources.some((s) => Number.isInteger(s.sourceId) && s.sourceId < 145);
+    this.is2014 = this.race.sources.some((s) => {
+      const force2014 = DICTIONARY.source.is2014.includes(s.sourceId);
+      if (force2014) return true;
+      const force2024 = DICTIONARY.source.is2024.includes(s.sourceId);
+      if (force2024) return false;
+      return Number.isInteger(s.sourceId) && s.sourceId < 145;
+    });
 
+    this.isLegacy = this.race.isLegacy;
     this.#fixups();
     this.compendiumRacialTraits = compendiumRacialTraits;
     this._generateDataStub();
@@ -106,7 +112,7 @@ export default class DDBRace {
     this.isSubRace = this.race.isSubRace || this.groupName !== this.fullName;
 
     const sourceIds = this.race.sources.map((sm) => sm.sourceId);
-    this.legacy = CONFIG.DDB.sources.some((ddbSource) =>
+    this.legacy = this.race.isLegacy || CONFIG.DDB.sources.some((ddbSource) =>
       sourceIds.includes(ddbSource.id)
       && DICTIONARY.sourceCategories.legacy.includes(ddbSource.sourceCategoryId),
     );